Update bytewise/bytewise
Adds complete (not fully tested) implementation of
writers to bytewise. Integrates writers with varints
and buffers.

Due to a bug in clang (which should be fixed in 7.0.0)
fold expression in writer's >> operator cannot be
used. Rolled back to more primitive template
recursion. Fix when a new version of the compiler is
released.