Add crypto/keyexchanger
Basic implementation of key exchanger. Method
exchange() returns a master key from which
receive and transmit secretbox symmetric keys can
be generated. It automatically takes care of
determining which key should be assigned to which
end of the connection, using a tie-breaker based on
the oreder between the public keys involved.