Update async/promise
Awaitable interface seems to be working. Instead
of making a copy of the value upon return, it would
be probably more efficient to add the coroutine
awakening to the arc. It would also be good to
directly capture the coroutine handle with a
closure, but it can't be taken by reference and
the resume method is not marked const.