Update async/promise
Found peculiar vulnerability: a promise deleting
itself in the callback produces a segmentation fault.
In order for the arc to always survive until the end
of the resolve / reject, a backup of its shared_ptr
is made in the stack. This way, even if the promise
is deleted, the reference counter never reaches 0 and
the arc is preserved until the resolve / reject
function returns.