Update async/promise
Complete refactor of the promise awaitable interface.
Instead of trying to maintain consistency in
reference counting, a promise no longer keeps an arc
pointer in its coroutine structure.

Since when the callback set in await_suspend is
called the coroutine will immediately call
await_resume in a higher level of the stack, a pointer
can be held in the promise to the value resulting
from the promise without any significant expense due
to data copy.

Removing the arc pointer from the promise coroutine
structure allows the promise object to become
completely inert with respect to the promise: if
the promise is successfully resolved, then a
coroutine can successfully resume from that.