Update drop/async/promise
Critical bugfix: in the awaitable interface of a
promise, whenever await_ready returned true, no
call to then / except was ever issued on the promise.

However, if the promise was rejected, this would
result in its destructor being called without any
error handler: this is due to the fact that the
error was actually handled by the coroutine itself
in await_resume.

Now, before await_ready is called, a dummy value and
error handler are added to the promise. This can
probably be optimized, but it is not critical for
the moment.