Update poseidon/brahms/brahms
Refactors brahms to use std :: array rather than
C-style arrays. This should be done throughout the
entire project, but it is not a critically urgent
task.

Also replaces the internally constructed, hard-wired
dialer with a typedef-ed parameter in the settings,
that can be easily changed when the deploy context
changes.