Update poseidon/brahms/brahms
The issue seems to be fixed by swapping the mutex for
a recursive_mutex in brahms. This is due to the fact
that the crontab thread, when running on local
sockets, might go through a whole pull request
without ever suspending the coroutine, thus resulting
in the same mutex being locked twice (once in run,
and once in pull).