Update drop/distributed/gossiper
When looping on connections to broadcast a gossip, a
gossiper might encounter a connection that was closed
in the meantime. This would result in calling the
on <close> handler, which would call drop, which in
turn would remove the connection from the sessions.
This results in segmentation faults, since it results
in removing elements from an unordered_map while still
looping on it.

Refactors the removal procedure as a lazy garbage
collection mechanism: drop() actually sets an optional
session to null, and it will be removed only the next
time a message is broadcasted.